requirements.txt
flask
flask-cors
python-docx
gunicorn
requests